🏗️ Architect: Modularized observations report logic
Decomposed the `apts/observations/html.py` God Mixin into a modular `apts/observations/report/` package. Logic is separated into specialized mixins: `ReportTemplateMixIn` (template management and CSS injection in `templates.py`) and `ReportDataMixIn` (data preparation and HTML-safe sanitization in `data.py`). The public API `HtmlExportMixIn` is recomposed in `base.py` and re-exported via `__init__.py` for backward compatibility. This refactor follows the Single Responsibility Principle by decoupling reporting data preparation from HTML rendering and template management.
