+++ apts/observations/report/__init__.py
@@ -1,0 +1,3 @@
+from .base import HtmlExportMixIn
+
+__all__ = ["HtmlExportMixIn"]
